# docs/openapi-integral-v8.yaml
openapi: 3.1.0
...
info:
  title: INTEGRAL — API Engine (Audit AI v8)
  version: 0.1.0
  description: >-
    Kontrak REST yang v7 ajukan untuk integrasi dengan SIMWAS v2 (INTEGRAL).
    SIMWAS memanggil v7 per-tahap; v7 mengembalikan data terstruktur (JSON) dan
    artefak dokumen (.docx). Dokumen ini mencerminkan endpoint nyata di backend
    FastAPI v7 plus beberapa endpoint `proposed` yang masih perlu dibangun.
servers:
  - url: http://localhost:8000
    description: Dev lokal (Docker)
  - url: http://127.0.0.1:8000
    description: >-
      Produksi yang direkomendasikan — v7 dengar di loopback / jaringan privat
      pada server yang sama dengan SIMWAS; TIDAK diekspos ke publik.
...
security:
  - SimwasSSO: []
components:
  securitySchemes:
    SimwasSSO:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        [proposed] Identitas user. Produksi: JWT bertanda-tangan SIMWAS,
        diverifikasi v7 lewat JWKS publik SIMWAS (klaim: sub, role AT|KT|PT|PM,
        nama, exp). Saat ini: JWT internal v7 (HS256) dari POST /auth/login,
        dikirim sebagai `Authorization: Bearer <token>` atau query `?_token=`
        (fallback khusus EventSource/SSE).
    SimwasApiKey:
      type: apiKey
      in: header
      name: X-SIMWAS-API-KEY
      description: >-
        [proposed] Auth server-to-server untuk panggilan SIMWAS → v7 yang tidak
        membawa konteks user (mis. sync ST). Mengikuti pola yang sudah ada di
        v7 (cacm_agent_api_key / X-API-Key).
  parameters:
    PenugasanId:
      name: penugasan_id
      in: path
      required: true
      schema: { type: integer }
    FilePath:
      name: path
      in: query
      required: true
      description: Path relatif terhadap folder penugasan (validasi anti-traversal).
      schema: { type: string, example: "_KKP/temuan.json" }
  schemas:
...
    PenugasanStatus:
      type: string
      enum: [USULAN_CACM, DRAFT, INGESTING, KKP_IN_PROGRESS, KKP_QC, KKP_DONE, LHP_IN_PROGRESS, LHP_QC, LHP_DONE]
      description: Diturunkan backend dari artefak di disk — bukan kolom yang ditulis langsung.
...
    PenugasanCreate:
      type: object
      required: [obyek, skill]
      properties:
        obyek: { type: string, example: "Reviu RKA App PDP 2026" }
        skill: { type: string, example: "reviu-rka-kl", description: "Slug skill (lihat GET /skills)" }
        nomor_st: { type: string, nullable: true }
        tanggal_st: { type: string, format: date, nullable: true }
    PenugasanOut:
      type: object
      properties:
        id: { type: integer }
        kode: { type: string, example: "WF/01/IJ.3/06/2026" }
        obyek: { type: string }
        skill: { type: string }
        nomor_st: { type: string, nullable: true }
        tanggal_st: { type: string, format: date, nullable: true }
        status: { $ref: '#/components/schemas/PenugasanStatus' }
        folder_path: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    FileEntry:
      type: object
      properties:
        name: { type: string }
        path: { type: string }
        size: { type: integer }
        mtime: { type: number }
        ext: { type: string }
    FilesListing:
      type: object
      properties:
        penugasan_id: { type: integer }
        folder_path: { type: string }
        categories:
          type: array
          items:
            type: object
            properties:
              key: { type: string, example: "KKP" }
              label: { type: string }
              files:
                type: array
                items: { $ref: '#/components/schemas/FileEntry' }
...
paths:
...
  /penugasan:
    get:
      tags: [penugasan]
      summary: List penugasan (status diturunkan dari artefak)
      x-status: live
      responses:
        '200':
          description: Daftar
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/PenugasanOut' } }
    post:
      tags: [penugasan]
      summary: Buat penugasan (PT) — di produksi sebaiknya lewat /api/simwas/st/sync
      x-status: live
      x-role: PT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PenugasanCreate' }
      responses:
        '201':
          description: Dibuat
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PenugasanOut' }
...
  /dokumen:
    get:
      tags: [dokumen]
      summary: List dokumen penugasan
      x-status: live
      parameters:
        - name: penugasan_id
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Daftar dokumen }
    post:
      tags: [dokumen]
      summary: Upload dokumen bukti (AT) — auto-digest di background
      x-status: live
      x-role: AT
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [penugasan_id, file]
              properties:
                penugasan_id: { type: integer }
                jenis: { type: string, description: "TOR|RAB|PBJ|ST|SURVEY|OTHER" }
                file: { type: string, format: binary }
      responses:
        '201': { description: Terupload (status UPLOADED/INGESTING/READY) }
  /dokumen/{dokumen_id}:
    delete:
      tags: [dokumen]
      summary: Hapus dokumen + reset artefak turunan (AT)
      x-status: live
      x-role: AT
      parameters:
        - name: dokumen_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Terhapus }
...
  /penugasan/{penugasan_id}/files:
    parameters: [ { $ref: '#/components/parameters/PenugasanId' } ]
    get:
      tags: [output]
      summary: List artefak hasil (KKP, LHP, QC, dll) per kategori
      x-status: live
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FilesListing' }
  /penugasan/{penugasan_id}/files/download:
    parameters:
      - { $ref: '#/components/parameters/PenugasanId' }
      - { $ref: '#/components/parameters/FilePath' }
    get:
      tags: [output]
      summary: Unduh artefak (mis. LHP-SUBSTANSI.docx, KKP.docx)
      x-status: live
      responses:
        '200':
          description: File biner/teks
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
  /penugasan/{penugasan_id}/files/preview:
    parameters:
      - { $ref: '#/components/parameters/PenugasanId' }
      - { $ref: '#/components/parameters/FilePath' }
    get:
      tags: [output]
      summary: Preview artefak teks (.md/.json/.txt/.csv) — mis. _KKP/temuan.json
      x-status: live
      parameters:
        - { name: max_bytes, in: query, required: false, schema: { type: integer, default: 51200 } }
      responses:
        '200':
          description: Isi teks
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  size_bytes: { type: integer }
                  ext: { type: string }
                  truncated: { type: boolean }
                  content: { type: string }

# -----------------------------------------------------------------------------
# Webhook keluar (v7 → SIMWAS) — PROPOSED
# -----------------------------------------------------------------------------
...